requirements.txt
Appium-Python-Client==1.0.2
backcall==0.2.0
certifi==2020.12.5
charset-normalizer==2.0.4
colorama==0.4.4
cx-Freeze==6.6
cycler==0.10.0
cytoolz==0.11.0
decorator==4.4.2
docopt==0.6.2
future==0.18.2
gym==0.18.0
idna==3.2
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
jupyter-client==6.1.11
jupyter-core==4.7.1
locket==0.2.1
matplotlib==3.4.2
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
olefile==0.46
pandas==1.2.3
parso==0.8.1
pickleshare==0.7.5
Pillow==7.2.0
pipreqs==0.4.10
prompt-toolkit==3.0.16
pyglet==1.5.0
Pygments==2.8.1
pytz==2021.1
pywin32==300
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.26.0
scikit-image==0.18.1
selenium==3.141.
torch==1.5.0
torchvision==0.6.0
traitlets==5.0.5
urllib3==1.26.3
wcwidth==0.2.5
wincertstore==0.2
yarg==0.1.9
